+++ test/com/amplitude/api/SessionTest.java
@@ -57,7 +57,7 @@ public class SessionTest extends BaseTest {
     @Test
     public void testDefaultStartSession() {
         long timestamp = System.currentTimeMillis();
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // trackSessionEvents is false, no start_session event added
@@ -75,13 +75,13 @@ public class SessionTest extends BaseTest {
 
         // log 1st event, initialize first session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEventAsync("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEventAsync("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 1);
 
         // log 2nd event past timeout, verify new session started
         long timestamp2 = timestamp1 + sessionTimeoutMillis;
-        amplitude.logEventAsync("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEventAsync("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 2);
 
@@ -106,17 +106,17 @@ public class SessionTest extends BaseTest {
 
         // log 3 events all just within session expiration window, verify all in same session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEventAsync("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEventAsync("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 1);
 
         long timestamp2 = timestamp1 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEventAsync("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 2);
 
         long timestamp3 = timestamp2 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("test3", null, null, null, null, timestamp3, false);
+        amplitude.logEventAsync("test3", null, null, null, null, null, timestamp3, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 3);
 
@@ -143,7 +143,7 @@ public class SessionTest extends BaseTest {
         amplitude.trackSessionEvents(true);
 
         long timestamp = System.currentTimeMillis();
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // trackSessions is true, start_session event is added
@@ -168,7 +168,7 @@ public class SessionTest extends BaseTest {
         amplitude.trackSessionEvents(true);
 
         long timestamp = System.currentTimeMillis();
-        amplitude.logEvent("test", null, null, null, null, timestamp, false);
+        amplitude.logEvent("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
         // trackSessions is true, start_session event is added
         assertEquals(getUnsentEventCount(), 2);
@@ -199,14 +199,14 @@ public class SessionTest extends BaseTest {
 
         // log 1st event, initialize first session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEventAsync("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEventAsync("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         // trackSessions is true, start_session event is added
         assertEquals(getUnsentEventCount(), 2);
 
         // log 2nd event past timeout, verify new session started
         long timestamp2 = timestamp1 + sessionTimeoutMillis;
-        amplitude.logEventAsync("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEventAsync("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         // trackSessions is true, end_session and start_session events are added
         assertEquals(getUnsentEventCount(), 5);
@@ -258,14 +258,14 @@ public class SessionTest extends BaseTest {
 
         // log 1st event, initialize first session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEvent("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEvent("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         // trackSessions is true, start_session event is added
         assertEquals(getUnsentEventCount(), 2);
 
         // log 2nd event past timeout, verify new session started
         long timestamp2 = timestamp1 + sessionTimeoutMillis;
-        amplitude.logEvent("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEvent("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         // trackSessions is true, end_session and start_session events are added
         assertEquals(getUnsentEventCount(), 5);
@@ -318,18 +318,18 @@ public class SessionTest extends BaseTest {
 
         // log 3 events all just within session expiration window, verify all in same session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEventAsync("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEventAsync("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         // trackSessions is true, start_session event is added
         assertEquals(getUnsentEventCount(), 2);
 
         long timestamp2 = timestamp1 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEventAsync("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 3);
 
         long timestamp3 = timestamp2 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("test3", null, null, null, null, timestamp3, false);
+        amplitude.logEventAsync("test3", null, null, null, null, null, timestamp3, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 4);
 
@@ -369,18 +369,18 @@ public class SessionTest extends BaseTest {
 
         // log 3 events all just within session expiration window, verify all in same session
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEvent("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEvent("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         // trackSessions is true, start_session event is added
         assertEquals(getUnsentEventCount(), 2);
 
         long timestamp2 = timestamp1 + sessionTimeoutMillis - 1;
-        amplitude.logEvent("test2", null, null, null, null, timestamp2, false);
+        amplitude.logEvent("test2", null, null, null, null, null, timestamp2, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 3);
 
         long timestamp3 = timestamp2 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("test3", null, null, null, null, timestamp3, false);
+        amplitude.logEventAsync("test3", null, null, null, null, null, timestamp3, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 4);
 
@@ -755,7 +755,7 @@ public class SessionTest extends BaseTest {
         assertFalse(amplitude.isInForeground());
 
         // logging an event before onResume will force a session check
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         looper.runToEndOfTasks();
         assertEquals(amplitude.previousSessionId, timestamp);
         assertEquals(amplitude.lastEventId, 1);
@@ -793,7 +793,7 @@ public class SessionTest extends BaseTest {
         assertFalse(amplitude.isInForeground());
 
         // logging an event before onResume will force a session check
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         looper.runToEndOfTasks();
         assertEquals(amplitude.previousSessionId, timestamp);
         assertEquals(amplitude.lastEventId, 2);
@@ -852,19 +852,19 @@ public class SessionTest extends BaseTest {
         amplitude.setSessionTimeoutMillis(sessionTimeoutMillis);
 
         long timestamp1 = System.currentTimeMillis();
-        amplitude.logEventAsync("test1", null, null, null, null, timestamp1, false);
+        amplitude.logEventAsync("test1", null, null, null, null, null, timestamp1, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 1);
 
         // log out of session event just within session expiration window
         long timestamp2 = timestamp1 + sessionTimeoutMillis - 1;
-        amplitude.logEventAsync("outOfSession", null, null, null, null, timestamp2, true);
+        amplitude.logEventAsync("outOfSession", null, null, null, null, null, timestamp2, true);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 2);
 
         // out of session events do not extend session, 2nd event will start new session
         long timestamp3 = timestamp1 + sessionTimeoutMillis;
-        amplitude.logEventAsync("test2", null, null, null, null, timestamp3, false);
+        amplitude.logEventAsync("test2", null, null, null, null, null, timestamp3, false);
         looper.runToEndOfTasks();
         assertEquals(getUnsentEventCount(), 3);
 
@@ -893,7 +893,7 @@ public class SessionTest extends BaseTest {
         Robolectric.getForegroundThreadScheduler().advanceTo(1);
 
         // log an event, should not be uploaded
-        amplitude.logEventAsync("testEvent", null, null, null, null, timestamps[0], false);
+        amplitude.logEventAsync("testEvent", null, null, null, null, null, timestamps[0], false);
         looper.runOneTask();
         looper.runOneTask();
         assertEquals(getUnsentEventCount(), 1);
@@ -924,7 +924,7 @@ public class SessionTest extends BaseTest {
         Robolectric.getForegroundThreadScheduler().advanceTo(1);
 
         // log an event, should not be uploaded
-        amplitude.logEventAsync("testEvent", null, null, null, null, timestamps[0], false);
+        amplitude.logEventAsync("testEvent", null, null, null, null, null, timestamps[0], false);
         looper.runOneTask();
         assertEquals(getUnsentEventCount(), 1);
 
@@ -998,7 +998,7 @@ public class SessionTest extends BaseTest {
         amplitude.trackSessionEvents(true);
 
         long timestamp = System.currentTimeMillis();
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // trackSessions is true, start_session event is added
@@ -1054,7 +1054,7 @@ public class SessionTest extends BaseTest {
         amplitude.trackSessionEvents(true);
 
         long timestamp = System.currentTimeMillis();
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // trackSessions is true, start_session event is added
@@ -1092,7 +1092,7 @@ public class SessionTest extends BaseTest {
         amplitude.trackSessionEvents(false);
 
         long timestamp = System.currentTimeMillis();
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // trackSessions is false, there should only be 1 event
@@ -1113,7 +1113,7 @@ public class SessionTest extends BaseTest {
         assertEquals(session_event.optString("session_id"), String.valueOf(timestamp));
 
         // log an event with new user id and session
-        amplitude.logEventAsync("test", null, null, null, null, timestamp, false);
+        amplitude.logEventAsync("test", null, null, null, null, null, timestamp, false);
         Shadows.shadowOf(amplitude.logThread.getLooper()).runToEndOfTasks();
 
         // verify post set user id